pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "konkur1405-bot"
version = "1.0.0"
description = "ربات پیشرفته کنکور ۱۴۰۵ با قابلیت‌های جامع"
authors = [
    {name = "SansssAK", Channel="@video_amouzeshi", email = "ssanssak@gmail.com"},
]
dependencies = [
    "python-telegram-bot>=20.7",
    "sqlalchemy>=2.0.23",
    "python-jdatetime>=3.8.0",
    "flask>=2.3.3",
    "gunicorn>=21.2.0",
    "apscheduler>=3.10.4",
    "pytz>=2023.3",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.1.1",
    "numpy>=1.25.2",
    "cryptography>=41.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "black>=23.9.1",
    "flake8>=6.1.0",
]

[project.scripts]
konkur-bot = "app:main"

[tool.black]
line-length = 88
target-version = ['py311']
